+++ src/main/java/ru/dpopkov/tasktracker/Main.java
@@ -1,6 +1,10 @@
 package ru.dpopkov.tasktracker;
 
+import ru.dpopkov.tasktracker.repositories.inmemory.SimpleProjectRepository;
 import ru.dpopkov.tasktracker.repositories.inmemory.SimpleUserRepository;
+import ru.dpopkov.tasktracker.services.ProjectService;
+import ru.dpopkov.tasktracker.services.ProjectServiceImpl;
+import ru.dpopkov.tasktracker.services.UserService;
 import ru.dpopkov.tasktracker.services.UserServiceImpl;
 import ru.dpopkov.tasktracker.ui.*;
 import ru.dpopkov.tasktracker.ui.actions.Actions;
@@ -9,7 +13,9 @@ import ru.dpopkov.tasktracker.ui.actions.ActionPack;
 public class Main {
 
     public static void main(String[] args) {
-        TaskTracker taskTracker = new TaskTracker(new UserServiceImpl(new SimpleUserRepository()));
+        ProjectService projectService = new ProjectServiceImpl(new SimpleProjectRepository());
+        UserService userService = new UserServiceImpl(new SimpleUserRepository());
+        TaskTracker taskTracker = new TaskTracker(userService, projectService);
 
         UiInput input = new ConsoleUiInput();
         UiOutput output = new ConsoleUiOutput();