+++ src/main/java/ru/dpopkov/tasktracker/TaskTracker.java
@@ -1,6 +1,8 @@
 package ru.dpopkov.tasktracker;
 
+import ru.dpopkov.tasktracker.model.Project;
 import ru.dpopkov.tasktracker.model.User;
+import ru.dpopkov.tasktracker.services.ProjectService;
 import ru.dpopkov.tasktracker.services.UserService;
 
 import java.util.List;
@@ -9,9 +11,11 @@ import java.util.Optional;
 public class TaskTracker {
 
     private final UserService userService;
+    private final ProjectService projectService;
 
-    public TaskTracker(UserService userService) {
+    public TaskTracker(UserService userService, ProjectService projectService) {
         this.userService = userService;
+        this.projectService = projectService;
     }
 
     public void createUser(String firstName, String lastName) {
@@ -29,5 +33,22 @@ public class TaskTracker {
 
     public List<User> getAllUsers() {
         return userService.getAll();
+    }
+
+    public void createProject(String name, String description) {
+        Project project = new Project(name, description);
+        projectService.create(project);
+    }
+
+    public Optional<Project> getProjectById(int id) {
+        return projectService.findById((long) id);
+    }
+
+    public boolean deleteProject(int id) {
+        return projectService.delete((long) id);
+    }
+
+    public List<Project> getAllProjects() {
+        return projectService.getAll();
     }
 }