+++ src/main/java/ru/dpopkov/tasktracker/repositories/inmemory/SimpleProjectRepository.java
@@ -1,6 +1,7 @@
 package ru.dpopkov.tasktracker.repositories.inmemory;
 
 import ru.dpopkov.tasktracker.model.Project;
+import ru.dpopkov.tasktracker.repositories.ProjectRepository;
 
-public class SimpleProjectRepository extends SimpleBaseRepository<Project> {
+public class SimpleProjectRepository extends SimpleBaseRepository<Project> implements ProjectRepository {
 }