+++ src/main/java/ru/dpopkov/tasktracker/services/ProjectService.java
@@ -1,0 +1,6 @@
+package ru.dpopkov.tasktracker.services;
+
+import ru.dpopkov.tasktracker.model.Project;
+
+public interface ProjectService extends BaseService<Project> {
+}