+++ src/main/java/ru/dpopkov/tasktracker/services/ProjectServiceImpl.java
@@ -1,0 +1,11 @@
+package ru.dpopkov.tasktracker.services;
+
+import ru.dpopkov.tasktracker.model.Project;
+import ru.dpopkov.tasktracker.repositories.ProjectRepository;
+
+public class ProjectServiceImpl extends AbstractBaseService<Project> implements ProjectService {
+
+    public ProjectServiceImpl(ProjectRepository repository) {
+        super(repository);
+    }
+}