+++ src/main/java/ru/dpopkov/tasktracker/ui/actions/ActionPack.java
@@ -3,6 +3,7 @@ package ru.dpopkov.tasktracker.ui.actions;
 import ru.dpopkov.tasktracker.Action;
 import ru.dpopkov.tasktracker.ActionType;
 import ru.dpopkov.tasktracker.TaskTracker;
+import ru.dpopkov.tasktracker.model.Project;
 import ru.dpopkov.tasktracker.model.User;
 import ru.dpopkov.tasktracker.ui.Messages;
 import ru.dpopkov.tasktracker.ui.UiInput;
@@ -10,9 +11,13 @@ import ru.dpopkov.tasktracker.ui.UiOutput;
 
 import java.util.*;
 
+/**
+ * Contains pack of {@link Action} instances that are using shared input/output and {@link TaskTracker} instance.
+ */
 public class ActionPack implements Iterable<Action> {
 
     private final String enterUserId = Messages.INSTANCE.get("enter_user_id");
+    private final String enterProjectId = Messages.INSTANCE.get("enter_project_id");
     private final UiInput input;
     private final UiOutput output;
     private final TaskTracker tracker;
@@ -35,6 +40,10 @@ public class ActionPack implements Iterable<Action> {
         actions.add(new ShowUserByIdAction());
         actions.add(new DeleteUserAction());
         actions.add(new ShowAllUsersAction());
+        actions.add(new AddProjectAction());
+        actions.add(new FindProjectByIdAction());
+        actions.add(new DeleteProjectAction());
+        actions.add(new ShowAllProjectsAction());
     }
 
     private class AddUserAction extends BaseAction {
@@ -66,7 +75,7 @@ public class ActionPack implements Iterable<Action> {
 
         @Override
         public void execute() {
-            int userId = readUserId();
+            int userId = readId(enterUserId);
             Optional<User> result = getTracker().getUserById(userId);
             if (result.isPresent()) {
                 getOutput().print(format(result.get()));
@@ -87,7 +96,7 @@ public class ActionPack implements Iterable<Action> {
 
         @Override
         public void execute() {
-            int userId = readUserId();
+            int userId = readId(enterUserId);
             boolean result = getTracker().deleteUser(userId);
             if (result) {
                 getOutput().println(userDeleted);
@@ -107,18 +116,96 @@ public class ActionPack implements Iterable<Action> {
         public void execute() {
             final TaskTracker tracker = getTracker();
             for (User user : tracker.getAllUsers()) {
-                getOutput().print(format(user));
+                getOutput().println(format(user));
             }
         }
 
     }
 
-    private int readUserId() {
-        output.prompt(enterUserId);
+    private class AddProjectAction extends BaseAction {
+
+        private final String projectName = Messages.INSTANCE.get("enter_project_name");
+        private final String projectDescription = Messages.INSTANCE.get("enter_project_description");
+
+        public AddProjectAction() {
+            super(input, output, tracker, ActionType.ADD_PROJECT);
+        }
+
+        @Override
+        public void execute() {
+            getOutput().prompt(projectName);
+            String name = getInput().readString();
+            getOutput().prompt(projectDescription);
+            String description = getInput().readString();
+            getTracker().createProject(name, description);
+        }
+    }
+
+    private class FindProjectByIdAction extends BaseAction {
+        private final String cannotFind = Messages.INSTANCE.get("cannot_find_project_id");
+
+        protected FindProjectByIdAction() {
+            super(input, output, tracker, ActionType.FIND_PROJECT_BY_ID);
+        }
+
+        @Override
+        public void execute() {
+            int projectId = readId(enterProjectId);
+            Optional<Project> result = tracker.getProjectById(projectId);
+            if (result.isPresent()) {
+                getOutput().println(format(result.get()));
+            } else {
+                getOutput().println(cannotFind + " " + projectId);
+            }
+        }
+    }
+
+    private class DeleteProjectAction extends BaseAction {
+
+        private final String projectDeleted = Messages.INSTANCE.get("project_deleted");
+        private final String cannotFind = Messages.INSTANCE.get("cannot_find_project_id");
+
+        protected DeleteProjectAction() {
+            super(input, output, tracker, ActionType.DELETE_PROJECT);
+        }
+
+        @Override
+        public void execute() {
+            int projectId = readId(enterProjectId);
+            boolean deleted = tracker.deleteProject(projectId);
+            if (deleted) {
+                getOutput().println(projectDeleted);
+            } else {
+                getOutput().println(cannotFind + " " + projectId);
+            }
+        }
+    }
+
+    private class ShowAllProjectsAction extends BaseAction {
+
+        protected ShowAllProjectsAction() {
+            super(input, output, tracker, ActionType.SHOW_ALL_PROJECTS);
+        }
+
+        @Override
+        public void execute() {
+            List<Project> all = tracker.getAllProjects();
+            all.forEach(p -> getOutput().println(format(p)));
+        }
+    }
+
+    // todo: Simplify action classes -> remove input/output/tracker
+
+    private int readId(String idPrompt) {
+        output.prompt(idPrompt);
         return input.readInt();
     }
 
     private String format(User user) {
-        return String.format("%2d : %s : %s%n", user.getId(), user.getFirstName(), user.getLastName());
+        return String.format("%2d : %s : %s", user.getId(), user.getFirstName(), user.getLastName());
+    }
+
+    private String format(Project project) {
+        return String.format("%2d : %s : %s", project.getId(), project.getName(), project.getDescription());
     }
 }